Message to user if there are no expired time slots to clear

+++ backend/test/views.py
@@ -356,6 +356,8 @@ def clear_expired_time_slots(request):
         date__lte=today,
         endTime__lte=rightnow,
     )
+    if len(expiredres)==0:
+        return Response({'message':'No expired times lots to clear'})
     for res in expiredres:
         res.delete()
     return Response({'message':'Expired time slots have been cleared'})
